Rocky support (#17)
* Newton support

* Fix routing issue with 1804 Ubuntu VM

On Ubunu18.04, netplan is responsible for bringing up the networks
interfaces, and when it runs DHCP on multiple interfaces, it adds all
the routes for default gateway for each interface. Before this, the
default gateway entry was only added for primary interface. This is
causing routing issue on Ubuntu 1804 VM.

The fix will add explicit entry in network system config to ignore route
from secondary interface DHCP response.

From Rocky release of OpenStack Ubuntu 1804 VM will be used to bring a
AIO OpenStack.

* Rocky install files